+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Devop
 
+* hide provider with 0 rates [#3167](https://github.com/MyEtherWallet/MyEtherWallet/pull/3167)
+* fix console errors [#3168](https://github.com/MyEtherWallet/MyEtherWallet/pull/3168)
 * btc address check [#3159](https://github.com/MyEtherWallet/MyEtherWallet/pull/3159)
 * Fix swap provider list [#3152](https://github.com/MyEtherWallet/MyEtherWallet/pull/3152)
 * Add scrollable option to app modal [#3145](https://github.com/MyEtherWallet/MyEtherWallet/pull/3145)